Refactor app class
